Removed unneeded parts of waveform tests and added tests on priors
